refactor: create separate componenet for profile images
